V3
-> Done #76
-> Done #75
-> Done #74
-> Done #72
-> Done #71
-> Removed non used code
-> Added namespaces for telemetry and error handler
-> Code cleanup
-> Fix every SQL injection
-> Code formatting
-> Removed old code
-> Some more namespaces changes
-> Now using a SettingsManager
-> Fixed some problems and bugs
-> Fixed some more SQL injections
-> Finish #73
-> Code formatting / cleanup
-> Fixed more bugs
-> Fixed the login and register and discord error
-> Fixed the ticket error for reply
-> Fixed the connection close error for new tickets
-> Fix Undefined array key "ticketuuid"
-> Done #80
-> Done #79
-> Done #77
-> Finish #78
-> Fixed server creation page with no name
-> Done #84
-> Done #83
-> Done #81
-> Update README

+++ view/admin/api/delete.php
@@ -2,7 +2,7 @@
 include(__DIR__ . '/../../requirements/page.php');
 include(__DIR__ . '/../../requirements/admin.php');
 
-if (isset($_GET['id'])) {
+if (isset($_GET['id']) && !$_GET['id'] == "") {
     $keyid = mysqli_real_escape_string($conn,$_GET['id']);
     mysqli_query($conn, "DELETE FROM mythicaldash_apikeys WHERE `mythicaldash_apikeys`.`id` = " . $keyid . "");
     header('location: /admin/api');
